pyproject.toml
[project]
name = "quicksql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yevgeny Bulochnik", email = "yevgeny.bulochnik@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.3",
    "google-cloud-bigquery[pyarrow]>=3.25.0",
    "ibis-framework[duckdb]>=11.0.0",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "watchdog>=6.0.0",
    "textual>=0.75.0",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-watcher>=0.6.2",
]
